Add copy of Starter Workflows’s license
GitHub’s starter-workflows repo is available under the MIT License. The
MIT License says:

> The above copyright notice and this permission notice shall be included in all
> copies or substantial portions of the Software.

Before this commit, the just-the-docs-template repo did not include a
copy of the start-workflows repo’s version of the MIT License. It
contained a link to the starter-workflows repo’s LICNESE file, but a
link is not good enough. You need to include an actual copy of the
licnese and copyright notice [1].

Before this commit, we did actually include a copy of the MIT License
because this repo itself is available under the MIT License.
Unfortunately, the version of the MIT License that this repo uses is
different than the version of the MIT License that the starter-workflows
repo uses. Additionally, we did not include the copyright notice for the
starter-workflows repo at all before this commit.

This commit was cherry-picked from this PR [2].

[1]: <https://writing.kemitchell.com/2019/03/09/Deprecation-Notice.html#mit-and-bsd-breed-confusion>
[2]: <https://github.com/just-the-docs/just-the-docs-template/pull/50>