Create CONTRIBUTING.md
The main goal of this commit is to allow any employee to contribute to
the SmartKids Docs, even if all they’ve been given is a link to
<https://docs.smartkidsllc.com>. A previous commit [1] made
<https://docs.smartkidsllc.com> link to
<https://github.com/smartkidsllc/smartkidsllc.github.io>. This commit
makes it so that once a reader makes it to the GitHub repo, they can
simply start reading and clicking on links in order to figure out how to
contribute.

The documentation for contributing was written to not assume that
readers already know how to use GitHub. Not all SmartKids employees are
technically savy, and the ones who are technically savy don’t
necessarily know how to submit a GitHub pull request.

[1]: 8087675 (Rewrite homepage, 2024-08-02)